moved ViewTrace classes to their own package
