+++ src/traceView/trace-view-controller.js
@@ -1,6 +1,6 @@
 /* global ace */
 
-import {AceUtils} from "./ace-utils";
+import {AceUtils} from "../utils/ace-utils";
 import {TraceViewModel} from "./trace-view-model";
 import {TraceModel} from '../traceService/trace-model';
