+++ src/traceView/trace-view-model.js
